+++ services/file-upload/internal/interfaces/storage.go
@@ -1,4 +1,4 @@
-package clients
+package interfaces
 
 import (
 	"context"
